requirements.txt
Django==1.7
Pillow==2.5.3
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-geoposition==0.2.2
django-gmapify==0.1
django-leaflet==0.14.1
django-toolbelt==0.0.1
geolocation-python==0.1.0
gunicorn==19.1.1
psycopg2==2.5.4
six==1.8.0
static3==0.5.1
wsgiref==0.1.2
